[06b] optimize loop detection
instead of saving the order in which the guard visited the tiles on her
trail, we now register in which direction she left each tile previously.
As @krotlkpt pointed out, this allows for much simpler and more robust
and faster loop detection: as soon as the guard is about to leave a tile
in a direction she already took there earlier, she is stuck in a loop.

This optimization appears to bring down computation time for part II of
the day 6 challenge from more than half an hour to about a minute.